requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
